Now boost_program_options is used for parsing the command line.

+++ SDK/include/SDKFile.hpp
@@ -67,13 +67,13 @@ public:
     void replaceNewlineWithSpaces()
     {
         size_t pos = source_.find_first_of('\n', 0);
-        while(pos != -1)
+        while(pos != (size_t) -1)
         {
             source_.replace(pos, 1, " ");
             pos = source_.find_first_of('\n', pos + 1);
         }
         pos = source_.find_first_of('\r', 0);
-        while(pos != -1)
+        while(pos != (size_t) -1)
         {
             source_.replace(pos, 1, " ");
             pos = source_.find_first_of('\r', pos + 1);
